[Refactoring][Add]
Add the TGCN model. It’s a model that performs better on data where
there is a correlation between the spatial representation and the
temporality in the data. It's the combinaison of the GRU model and the
GCN model.
Give credit to the original authors of the T-GCN, on which the added
T-GCN model is based.

Rewrite the string docs of some classes and functions.

Signed-off-by: Guillaume COGONI <cogoni.guillaume@gmail.com>