Testing (#19)
* Add testing in the main.py
Update variable names to be more explicit

* Added the possibility of saving the plot

* Correct issue #15

* Change nodes params in utils_data.py
The param is the list of nodes to process not an
interger. So it should be a empty list as a default param.

* Add metrics table to metrics.py and a test.py
test.py is a file with a configurable to test saved model
the metrics table is export and saved as a json and txt.

* Added a test.py and export results in a txt file
The results of main.py and test.py are exported in txt file
The results of test.py will be exported as json
Results are updated in real time.