Testing (#23)
* Add testing in the main.py
Update variable names to be more explicit

* Added the possibility of saving the plot

* Correct issue #15

* Change nodes params in utils_data.py
The param is the list of nodes to process not an
interger. So it should be a empty list as a default param.

* Add metrics table to metrics.py and a test.py
test.py is a file with a configurable to test saved model
the metrics table is export and saved as a json and txt.

* Added a test.py and export results in a txt file
The results of main.py and test.py are exported in txt file
The results of test.py will be exported as json
Results are updated in real time.

* Add a percentage selection of the time serie.
You can set the parameter time_serie_percentage_length
to a percentage if you don't want to work with the full
time serie. The base parameter is 1.

* Training in federated will only save bestmodels
After each communication round and after each model
averaging, the mean_squared_error validation loss of each node
will be calculated to conserve only the best model.

* Add % of superior prediction between local and fed
Add a metric to calculate point par point superior
prediction  between the federated models and the local model.
Correct some unnecessary redundancy too in the code.

* Correct incorrect ploting in test.py

* Correct unnecessary print during table print.